fix: 🐛 select unique values in table
