+++ packages/parser/src/constants.ts
@@ -1,4 +1,8 @@
+import { Analysis } from './types';
+
 export const KEEN_VALUE = 'keen.value';
 export const KEEN_KEY = 'keen.key';
 
 export const DEFAULT_NAME = 'Result';
+
+export const USE_SEPARATOR_FOR_GROUPS: Analysis[] = ['select_unique'];
