+++ packages/parser/src/transformations/atomic-result/index.ts
@@ -1,0 +1,3 @@
+import { transformAtomicResult } from './atomic-result';
+
+export { transformAtomicResult };
