+++ packages/parser/src/transformations/extraction/index.ts
@@ -1,0 +1,3 @@
+import { transformExtraction } from './extraction';
+
+export { transformExtraction };
