+++ packages/parser/src/transformations/funnel/index.ts
@@ -1,0 +1,3 @@
+import { transformFunnel } from './funnel';
+
+export { transformFunnel };
