+++ packages/parser/src/transformations/intervals-array/index.ts
@@ -1,0 +1,3 @@
+import { transformIntervalsFromArray } from './intervals-array';
+
+export { transformIntervalsFromArray };
