+++ packages/parser/src/transformations/number/index.ts
@@ -1,0 +1,3 @@
+import { transformFromNumber } from './number';
+
+export { transformFromNumber };
