+++ packages/ui-core/src/components/table/table-cell.component.tsx
@@ -4,8 +4,10 @@ import { Container, StyledCell } from './table-cell.styles';
 
 import { Text } from '../../typography';
 
-import { CellValue } from './types';
+import { CellValue, CellTextAlignment } from './types';
 import { Typography } from '../../types';
+
+const VALUE_DELIMETER = ', ';
 
 type Props = {
   /** Cell typography */
@@ -20,15 +22,29 @@ type Props = {
 };
 
 const TableCell: FC<Props> = ({ typography, onClick, value }) => {
-  const textAlignment = useMemo(() => {
+  const { textAlignment, cellValue } = useMemo<{
+    textAlignment: CellTextAlignment;
+    cellValue: string | number | boolean | Date;
+  }>(() => {
     const valueType = typeof value;
-    return valueType === 'number' ? 'right' : 'left';
+    let cellValue;
+
+    if (Array.isArray(value)) {
+      cellValue = value.join(VALUE_DELIMETER);
+    } else {
+      cellValue = value;
+    }
+
+    return {
+      cellValue,
+      textAlignment: valueType === 'number' ? 'right' : 'left',
+    };
   }, [value]);
 
   return (
     <StyledCell data-testid="table-cell" onClick={e => onClick(e, value)}>
       <Container textAlignment={textAlignment}>
-        <Text {...typography}>{value}</Text>
+        <Text {...typography}>{cellValue}</Text>
       </Container>
     </StyledCell>
   );
