+++ packages/ui-core/src/components/table/types.ts
@@ -1,1 +1,3 @@
-export type CellValue = string | number | boolean | Date;
+export type CellValue = string | number | boolean | Date | string[] | number[];
+
+export type CellTextAlignment = 'left' | 'right';
